organized tests, note: excludes running tests that boot up server
