+++ src/test/java/org/launchcode/brewpub/controllers/HomeControllerTest.java
@@ -1,22 +1,44 @@
 package org.launchcode.brewpub.controllers;
 
-import org.junit.jupiter.api.Test;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.launchcode.brewpub.CustomUserDetailService;
+import org.launchcode.brewpub.models.data.BrewRepository;
+import org.launchcode.brewpub.models.data.PubRepository;
+import org.launchcode.brewpub.models.data.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.web.context.WebApplicationContext;
 
 import static org.hamcrest.Matchers.containsString;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@SpringBootTest
-@AutoConfigureMockMvc
+@RunWith(SpringRunner.class)
+@WebMvcTest(HomeController.class)
 public class HomeControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
+
+    @MockBean
+    private CustomUserDetailService customUserDetailService;
+
+    @Autowired
+    private WebApplicationContext context;
+
+    @MockBean
+    private UserRepository userRepository;
+
+    @MockBean
+    private BrewRepository brewRepository;
+
+    @MockBean
+    private PubRepository pubRepository;
 
     @Test
     public void homeControllerShouldReturnWelcomeMessage() throws Exception {
