+++ src/test/java/org/launchcode/brewpub/webAppTest/BrewPubApplicationTests.java
@@ -1,4 +1,4 @@
-package org.launchcode.brewpub;
+package org.launchcode.brewpub.webAppTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
