+++ src/test/java/org/launchcode/brewpub/webAppTest/TestingWebApplicationTest.java
@@ -1,4 +1,4 @@
-package org.launchcode.brewpub;
+package org.launchcode.brewpub.webAppTest;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
